TMBF: Fix a handful of small bugs
* set_adc_offset needs to ensure DDR is in long capture mode
* Fix tune measurement on backwards tune sweeps
* Fix magnitude computation in one extreme case for Tune PLL

git-svn-id: svn+ssh://serv2/home/subversion/repos/controls/diamond/trunk/ioc/TMBF@115006 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ tmbfApp/src/tune_follow.c
@@ -222,7 +222,11 @@ static void update_iq_angle_mag(void)
     /* Compute magnitude and angle. */
     current_angle = wrap_angle(
         180.0 / M_PI * atan2(current_q, current_i) - delay_offset_degrees);
-    current_magnitude = (int) sqrt(SQR(current_i) + SQR(current_q));
+    /* There is a tricky detail in this calculation.  When IQ saturates one of
+     * the possible limit values is (I,Q) = (-2^15,-2^15) with I^2+Q^2 = 2^31,
+     * which is too large for an int, hence the casts to unsigned int. */
+    current_magnitude = (int) sqrt(
+        (unsigned int) SQR(current_i) + (unsigned int) SQR(current_q));
 
     delta_angle = wrap_angle(current_angle - target_phase);
 }
